undo autoreload mixin
